feat(559): move common to common
move the common bits of code to the common package
