+++ packages/appeals-service-api/package.json
@@ -23,7 +23,6 @@
 		"@azure/app-configuration": "^1.3.1",
 		"@azure/identity": "3.0.1",
 		"@azure/service-bus": "^7.7.3",
-		"@azure/storage-blob": "^12.16.0",
 		"@babel/core": "^7.20.12",
 		"@cloudamqp/amqp-client": "^2.1.0",
 		"@pins/business-rules": "file:../business-rules",
