+++ packages/document-service-api/src/lib/blobStorage.js
@@ -1,29 +1,7 @@
-const { BlobServiceClient } = require('@azure/storage-blob');
 const fs = require('fs');
 const path = require('path');
 const logger = require('./logger');
 const config = require('../configuration/config');
-
-const initContainerClient = async () => {
-	logger.info('Connecting to blob storage');
-
-	const { connectionString, container } = config.storage;
-
-	let containerClient;
-
-	try {
-		const blobClient = BlobServiceClient.fromConnectionString(connectionString);
-		containerClient = blobClient.getContainerClient(container);
-		logger.info({ container }, 'Creating container if not present');
-		await containerClient.createIfNotExists();
-		logger.info('Successfully connected to blob storage');
-	} catch (err) {
-		logger.error({ err }, 'Failed to connect to blob storage');
-		throw err;
-	}
-
-	return containerClient;
-};
 
 const downloadFile = (containerClient, location) => {
 	try {
@@ -120,7 +98,6 @@ const saveMetadata = async (containerClient, metadata) => {
 };
 
 module.exports = {
-	initContainerClient,
 	downloadFile,
 	uploadFile,
 	deleteFile,
