+++ packages/document-service-api/src/lib/blobStorage.test.js
@@ -1,7 +1,5 @@
-const { BlobServiceClient } = require('@azure/storage-blob');
 const path = require('path');
 const {
-	initContainerClient,
 	downloadFile,
 	uploadFile,
 	deleteFile,
@@ -36,19 +34,6 @@ const mockGetContainerClient = {
 
 let containerClient;
 
-jest.mock('../configuration/config', () => ({
-	fileUpload: {
-		path: ''
-	},
-	logger: {
-		level: 'info'
-	},
-	storage: {
-		connectionString: '',
-		container: ''
-	}
-}));
-
 jest.mock('@azure/storage-blob', () => ({
 	BlobServiceClient: {
 		fromConnectionString: () => ({
@@ -70,27 +55,6 @@ describe('lib/blobStorage', () => {
 			}),
 			listBlobsFlat: jest.fn().mockReturnValue(multipleFilesReturnValue)
 		};
-	});
-
-	describe('initContainerClient', () => {
-		it('should return containerClient when the connection is initialised successfully', async () => {
-			const result = await initContainerClient();
-
-			expect(result).toStrictEqual(mockGetContainerClient);
-		});
-
-		it('should throw an error when an error occurs', async () => {
-			BlobServiceClient.fromConnectionString = jest.fn().mockImplementation(() => {
-				throw new Error('Internal Server Error');
-			});
-
-			try {
-				await initContainerClient();
-				throw new Error('Expected error not thrown');
-			} catch (err) {
-				expect(err.message).toEqual('Internal Server Error');
-			}
-		});
 	});
 
 	describe('downloadFile', () => {
